refactor: extract i18n directive handlers
